requirements.txt
flask==3.0.3
langchain==0.3.6
langchain-openai==0.2.5
langgraph==0.2.42
pydantic==2.9.2
cdp-sdk==0.10.3
python-dotenv==1.0.1
cdp-langchain==0.0.4
cdp-agentkit-core==0.0.3
flask-cors==5.0.0
gunicorn==20.1.0
web3==6.15.1
uniswap-python==0.7.1
eth-abi==4.2.1
eth-account==0.10.0
eth-typing==3.5.2
eth-utils==2.3.1